+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Section and Section::FlexiRecord now uses active\_attr
   * Role now uses active\_attr
   * RegisterField now uses active\_attr
+  * RegisterData now uses active\_attr
 
 ## Version 0.0.17